code/requirements.txt
beautifulsoup4==4.11.1
Flask==2.1.2
jieba==0.42.1
matplotlib==3.5.1
numpy==1.22.3
opencv_python==4.5.5.64
openpyxl==3.0.10
pandas==1.4.2
Pillow==9.1.1
pymysql==1.0.2
requests==2.27.1
torch==1.11.0
torchvision==0.12.0
xlwt==1.3.0
